update cpp structure in the R package
