--- R-package/src/Data.cpp
@@ -1,5 +1,0 @@
-//
-// Created by Mamba on 2020/2/18.
-//
-
-#include "Data.h"
