--- R-package/src/Metric.cpp
@@ -1,5 +1,0 @@
-//
-// Created by Mamba on 2020/2/18.
-//
-
-#include "Metric.h"
