--- R-package/src/model_fit.cpp
@@ -1,1 +1,0 @@
-
