+++ R-package/src/normalize.cpp
@@ -1,5 +1,5 @@
 //
-// Created by jk on 2020/3/8.
+// Created by Kangkang Jiang on 2020/3/8.
 //
 // #define R_BUILD
 #ifdef R_BUILD
@@ -47,7 +47,6 @@ void Normalize(Eigen::MatrixXd &X, Eigen::VectorXd &y, Eigen::VectorXd &weights,
 
 void Normalize(Eigen::MatrixXd &X, Eigen::MatrixXd &y, Eigen::VectorXd &weights, Eigen::VectorXd &meanx, Eigen::VectorXd &meany, Eigen::VectorXd &normx)
 {
-    // cout << "multigaussian normal" << endl;
     int n = X.rows();
     int p = X.cols();
     Eigen::VectorXd tmp(n);
@@ -56,7 +55,6 @@ void Normalize(Eigen::MatrixXd &X, Eigen::MatrixXd &y, Eigen::VectorXd &weights,
         meanx(i) = weights.dot(X.col(i)) / double(n);
     }
     meany = y.transpose() * weights / double(n);
-    // cout << "meany: " << meany << endl;
     for (int i = 0; i < p; i++)
     {
         X.col(i) = X.col(i).array() - meanx(i);
