--- R-package/src/screening.cpp
@@ -1,1 +1,0 @@
-
